+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Non-breaking changes:
+    - WOM-454: Integrate weather station into DiSQ.
 - Bug fixes:
     - WOM-518: Rectify static tracking offsets input range limits.
     - WOM-424: Simplified main dependencies and limited Python version to 3.10 and 3.11.